[FIX] ResNet-18 accuracy 68.934 -> 69.752
The problem was with resize step in the preprocessing. The problem pretty much described in resize docs : "The output image might be different depending on its type: when downsampling, the interpolation of PIL images and tensors is slightly different, because PIL applies antialiasing. This may lead to significant differences in the performance of a network".